pyproject.toml
[tool.poetry]
name = "library-catalog"
version = "0.1.0"
description = "Book Catalog Management Project"
authors = ["Svetlana Suvorova <svetlya4ok1999@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^1.4.42"
asyncpg = "^0.30.0"
databases = "^0.7.0"
pydantic = "^2.0"
pydantic-settings = '^2.2.1'
httpx = "^0.24.1"
loguru = "^0.7.0"
alembic = "^1.11.1"
python-dotenv = '^1.0'
psycopg2-binary = '^2.9'

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-bugbear = "^23.9.11"
pytest = "^7.4.0"

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true